Consolidate event scanner
